Address Claude review feedback
- Freeze every camera interaction during a gesture, not just dragPan,
  boxZoom and doubleClickZoom. Vertices are recorded in screen space and
  unprojected once at finish(), so a scroll-wheel zoom or arrow-key pan
  between two polygon clicks moved the earlier vertices over the map.
- Treat a stray click on a drag shape as a no-op. polygonFromPoints() had
  started returning null for it, but finish() then applied an empty match
  under mode "new", wiping the selection instead of leaving it alone.
- Hoist applyMatchedSelection into @geolibre/core so the map gestures and
  the layer-panel dialogs share one copy of the combine rule; the app module
  re-exports it for its existing importers.
